commitn

+++ produit.h
@@ -38,7 +38,7 @@ void setdate_expiration(QString);
 // cRuds
    bool ajouter();
    QSqlQueryModel    *   afficher();
-   bool modifier(float);
+   bool modifier(int);
    bool supprimer(int);
 
 private:
